:recycle: Move OpenAIServingModels to separate file
